Add 'wait' parameter to PowerShellCommand plugin
This update introduces a new 'wait' option to the PowerShellCommand plugin, allowing users to control whether to wait for a script to finish execution or not. Any value other than 'true' or 'yes' will be considered 'false'. The parameter has been fully implemented and validated to ensure the commands execute correctly.